+++ x/token/types/expected_keepers.go
@@ -1,8 +1,17 @@
 package types
 
-import sdk "github.com/cosmos/cosmos-sdk/types"
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/x/supply/exported"
+)
 
 // BankKeeper defines the expected bank Keeper
 type BankKeeper interface {
 	AddCoins(ctx sdk.Context, addr sdk.AccAddress, amt sdk.Coins) (sdk.Coins, sdk.Error)
+}
+
+// SupplyKeeper defines the expected bank Keeper
+type SupplyKeeper interface {
+	GetSupply(ctx sdk.Context) (supply exported.SupplyI)
+	SetSupply(ctx sdk.Context, supply exported.SupplyI)
 }